move: bean 클래스 각 새 패키지로 이동
repositories, services
